requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
Pillow>=9.0,<11.0
django-cors-headers>=4.0.0

scikit-learn
pandas
numpy
joblib
Pillow
rasterio
matplotlib
